Moved Modules to separate class instead of enum

+++ src/I-Synergy.Framework.Backend/Controllers/Accounts/BaseAccountController.cs
@@ -115,7 +115,7 @@ namespace ISynergy.Controllers.Base
                     LicenseName = model.Licensename,
                     Email = model.Email,
                     Password = model.Password,
-                    Modules = (int)model.Modules,
+                    Modules = model.Modules,
                     UsersAllowed = model.Users
                 })
                 .ConfigureAwait(false))
